fix(virtual_array): fix colorize crash, add memory tracking & unit tests
- Fix out-of-bounds crash when colorizing 3-channel VirtualTexture
- Fix channel offset in mix_normal_map for variable channel counts
- Standardize TileRegion halo type to glm::ivec4
- Eliminate per-pixel heap allocation bottleneck in for_each_pixel
- Add live_tile_count and live_memory_bytes across TileStorage, VirtualArray, and VirtualTexture
- Add comprehensive GoogleTest suites for VirtualArray and VirtualTexture

+++ HighMap/include/highmap/virtual_array/tile_region.hpp
@@ -45,7 +45,7 @@ struct TileRegion
   TileRegion(const TileKey    &key,
              const glm::vec4  &bbox,
              const glm::ivec2 &shape,
-             const glm::vec4  &halo = {0, 0, 0, 0});
+             const glm::ivec4 &halo = {0, 0, 0, 0});
 
   glm::vec2   cell_center(int i, int j) const;
   glm::vec2   cell_corner(int i, int j) const;
